test(B5): ASan+UBSan sanitizer leg + fix FB-009 runtime pointer UB
New `fxtest.py sanitize` builds & runs test_all (which imports the T5 suites)
under -fsanitize=address,undefined; wired into CI nightly.

Running it surfaced FB-009: fx_next_slice (array.impl.h) advanced a pointer by a
size_t `step` that is actually a signed byte offset (negative for reverse/
negative-step slices), which UBSan flags as unsigned pointer overflow. Cast both
advance sites to (ptrdiff_t). test_all (+ T5) now runs clean under ASan+UBSan.

Recorded in found_bugs.md, incl. the finding that ASan does NOT reliably catch
FB-005 (the wrap over-read lands in valid adjacent heap -- why it was silent),
so the reference-checked T5 suite remains the layer that catches it.

+++ runtime/ficus/impl/array.impl.h
@@ -405,14 +405,17 @@ FX_INLINE bool fx_next_slice(fx_arriter_pos_t* stack, int ndims)
     fx_arriter_pos_t* s = &stack[ndims-1];
     int i;
     if (++s->pos < s->size) {
-        s->ptr += s->step;
+        // step is size_t but holds a signed byte offset (negative for
+        // reverse/negative-step slices); use ptrdiff_t so the pointer
+        // arithmetic is well-defined instead of overflowing (UBSan).
+        s->ptr += (ptrdiff_t)s->step;
         return true;
     }
     for (i = ndims-2; i >= 0 && ++stack[i].pos >= stack[i].size; i--)
         ;
     if (i < 0)
         return false;
-    stack[i].ptr += stack[i].step;
+    stack[i].ptr += (ptrdiff_t)stack[i].step;
     for (++i; i < ndims; i++) {
         stack[i].ptr = stack[i-1].ptr;
         stack[i].pos = 0;
